+++ library/src/main/java/cloud/pace/sdk/appkit/app/api/AppAPI.kt
@@ -1,8 +1,6 @@
 package cloud.pace.sdk.appkit.app.api
 
 import cloud.pace.sdk.api.API
-import cloud.pace.sdk.api.geo.GeoAPIClient
-import cloud.pace.sdk.api.geo.GeoAPIResponse
 import cloud.pace.sdk.api.poi.POIAPI.apps
 import cloud.pace.sdk.api.poi.generated.model.LocationBasedApp
 import cloud.pace.sdk.api.poi.generated.model.LocationBasedApps
@@ -10,6 +8,8 @@ import cloud.pace.sdk.api.poi.generated.model.LocationBasedAppsWithRefs
 import cloud.pace.sdk.api.poi.generated.request.apps.CheckForPaceAppAPI.checkForPaceApp
 import cloud.pace.sdk.api.poi.generated.request.apps.GetAppAPI.getApp
 import cloud.pace.sdk.api.poi.generated.request.apps.GetAppsAPI.getApps
+import cloud.pace.sdk.poikit.geo.GeoAPIClient
+import cloud.pace.sdk.poikit.geo.GeoAPIResponse
 import cloud.pace.sdk.poikit.utils.ApiException
 import cloud.pace.sdk.utils.enqueue
 import retrofit2.Call