+++ library/src/main/java/cloud/pace/sdk/utils/KoinConfig.kt
@@ -3,7 +3,6 @@ package cloud.pace.sdk.utils
 import android.content.Context
 import androidx.preference.PreferenceManager
 import androidx.room.Room
-import cloud.pace.sdk.api.geo.GeoAPIClient
 import cloud.pace.sdk.appkit.AppManager
 import cloud.pace.sdk.appkit.app.AppFragmentViewModel
 import cloud.pace.sdk.appkit.app.AppFragmentViewModelImpl
@@ -16,8 +15,6 @@ import cloud.pace.sdk.appkit.communication.AppEventManager
 import cloud.pace.sdk.appkit.communication.AppEventManagerImpl
 import cloud.pace.sdk.appkit.communication.AppModel
 import cloud.pace.sdk.appkit.communication.AppModelImpl
-import cloud.pace.sdk.appkit.geo.GeoAPIManager
-import cloud.pace.sdk.appkit.geo.GeoAPIManagerImpl
 import cloud.pace.sdk.appkit.geofences.GeofenceCallback
 import cloud.pace.sdk.appkit.network.NetworkChangeListener
 import cloud.pace.sdk.appkit.network.NetworkChangeListenerImpl
@@ -30,6 +27,9 @@ import cloud.pace.sdk.appkit.persistence.SharedPreferencesModel
 import cloud.pace.sdk.idkit.authorization.AuthorizationManager
 import cloud.pace.sdk.idkit.credentials.CredentialsManager
 import cloud.pace.sdk.poikit.database.POIKitDatabase
+import cloud.pace.sdk.poikit.geo.GeoAPIClient
+import cloud.pace.sdk.poikit.geo.GeoAPIManager
+import cloud.pace.sdk.poikit.geo.GeoAPIManagerImpl
 import cloud.pace.sdk.poikit.poi.download.TileDownloader
 import cloud.pace.sdk.poikit.routing.NavigationApiClient
 import cloud.pace.sdk.poikit.search.AddressSearchClient