+++ library/src/test/java/cloud/pace/sdk/appkit/AppManagerTest.kt
@@ -4,28 +4,19 @@ import android.content.Context
 import android.location.Location
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import cloud.pace.sdk.PACECloudSDK
-import cloud.pace.sdk.api.geo.GeoAPIFeature
-import cloud.pace.sdk.api.geo.GeoGasStation
-import cloud.pace.sdk.api.geo.Polygon
-import cloud.pace.sdk.appkit.app.api.AppAPI
 import cloud.pace.sdk.appkit.app.api.AppRepository
-import cloud.pace.sdk.appkit.app.api.AppRepositoryImpl
-import cloud.pace.sdk.appkit.app.api.UriManager
 import cloud.pace.sdk.appkit.communication.AppEventManager
 import cloud.pace.sdk.appkit.communication.AppModel
 import cloud.pace.sdk.appkit.communication.AppModelImpl
-import cloud.pace.sdk.appkit.geo.GeoAPIManager
 import cloud.pace.sdk.appkit.model.App
-import cloud.pace.sdk.appkit.persistence.CacheModel
 import cloud.pace.sdk.appkit.persistence.SharedPreferencesModel
 import cloud.pace.sdk.appkit.utils.*
 import cloud.pace.sdk.utils.*
 import io.mockk.every
 import io.mockk.mockk
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.runBlocking
 import org.junit.After
-import org.junit.Assert.*
+import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
@@ -391,111 +382,6 @@ class AppManagerTest : CloudSDKKoinComponent {
         val response3 = future.get()[2]
         assertEquals(app3.name, response3.name)
         assertEquals(app3.description, response3.description)
-    }
-
-    @Test
-    fun `poi id is in range`() = runBlocking {
-        val id = "e3211b77-03f0-4d49-83aa-4adaa46d95ae"
-
-        val geoApiManager = object : TestGeoAPIManager() {
-            override fun apps(latitude: Double, longitude: Double, completion: (Result<List<GeoGasStation>>) -> Unit) {
-                completion(Result.success(listOf(GeoGasStation(id, listOf("https://app.test")))))
-            }
-
-            override fun features(poiId: String, latitude: Double, longitude: Double, completion: (Result<List<GeoAPIFeature>>) -> Unit) {
-                completion(Result.success(listOf(getGeoAPIFeature(id))))
-            }
-        }
-
-        val testModule = module {
-            single<CacheModel> {
-                TestCacheModel()
-            }
-
-            single<AppAPI> {
-                TestAppAPI()
-            }
-
-            single<UriManager> {
-                TestUriUtils()
-            }
-
-            single<LocationProvider> {
-                TestLocationProvider(mockLocation)
-            }
-
-            single<GeoAPIManager> {
-                geoApiManager
-            }
-
-            single<AppRepository> {
-                AppRepositoryImpl(mockContext, get(), get(), get(), get())
-            }
-        }
-
-        setupKoinForTests(testModule)
-
-        assertTrue(appManager.isPoiInRange(id))
-    }
-
-    @Test
-    fun `poi id is not in range`() = runBlocking {
-        val id1 = "e3211b77-03f0-4d49-83aa-4adaa46d95ae"
-        val id2 = "992b77b6-5982-4848-88fe-ae2633308279"
-
-        val geoApiManager = object : TestGeoAPIManager() {
-            override fun apps(latitude: Double, longitude: Double, completion: (Result<List<GeoGasStation>>) -> Unit) {
-                completion(Result.success(listOf(GeoGasStation(id1, listOf("https://app.test")))))
-            }
-
-            override fun features(poiId: String, latitude: Double, longitude: Double, completion: (Result<List<GeoAPIFeature>>) -> Unit) {
-                completion(Result.success(listOf(getGeoAPIFeature(id2))))
-            }
-        }
-
-        val testModule = module {
-            single<CacheModel> {
-                TestCacheModel()
-            }
-
-            single<AppAPI> {
-                TestAppAPI()
-            }
-
-            single<UriManager> {
-                TestUriUtils()
-            }
-
-            single<LocationProvider> {
-                TestLocationProvider(mockLocation)
-            }
-
-            single<GeoAPIManager> {
-                geoApiManager
-            }
-
-            single<AppRepository> {
-                AppRepositoryImpl(mockContext, get(), get(), get(), get())
-            }
-        }
-
-        setupKoinForTests(testModule)
-
-        assertFalse(appManager.isPoiInRange(id1))
-    }
-
-    private fun getGeoAPIFeature(id: String): GeoAPIFeature {
-        return GeoAPIFeature(
-            id,
-            "Feature",
-            Polygon(listOf(polygon)),
-            mapOf(
-                "apps" to listOf(
-                    "type" to "fueling",
-                    "url" to "https://fueling.app.test"
-                )
-            )
-        )
     }
 
     private fun setupKoinForTests(module: Module) {